feat(reaction) checked reaction endpoint

+++ backend/AAiT-backend-group-13/usecase/reaction/command/update_handler.go
@@ -1,6 +1,7 @@
 package reactioncmd
 
 import (
+	er "github.com/group13/blog/domain/errors"
 	"github.com/group13/blog/domain/models"
 	icmd "github.com/group13/blog/usecase/common/cqrs/command"
 	irepo "github.com/group13/blog/usecase/common/i_repo"
@@ -25,37 +26,54 @@ func NewUpdateHandler(reactionRepo irepo.Reaction, blogRepo irepo.Blog) *UpdateH
 
 // Handle processes the command to update a reaction and adjusts the blog's reaction counts accordingly.
 func (h *UpdateHandler) Handle(cmd *UpdateCommand) (bool, error) {
-	// Find and delete existing reaction by user and blog ID
+
 	existingReaction, err := h.reactionRepo.FindReactionByUserIdAndBlogId(cmd.UserId, cmd.BlogId)
+
 	if err == nil && existingReaction != nil {
+
+		if cmd.IsLike == existingReaction.IsLike() {
+			return false, er.NewConflict("already liked the blog")
+		}
+
 		if err = h.reactionRepo.Delete(existingReaction.ID()); err != nil {
 			return false, err
 		}
 
-		// Adjust blog reaction count based on the old reaction
 		blog, err := h.blogRepo.GetSingle(cmd.BlogId)
+
 		if err != nil {
 			return false, err
 		}
 
 		if existingReaction.IsLike() {
-			if err = blog.UpdateLikeCount(false); err != nil {
+			err := blog.UpdateLikeCount(false)
+
+			if err != nil {
 				return false, err
 			}
-		} else {
-			if err = blog.UpdateDislikeCount(false); err != nil {
-				return false, err
-			}
+
+			h.blogRepo.Save(blog)
 		}
+
+		if !existingReaction.IsLike() {
+			err := blog.UpdateDislikeCount(false)
+
+			if err != nil {
+				return false, err
+			}
+
+			h.blogRepo.Save(blog)
+		}
+
 	}
 
-	// Create and save the new reaction
 	newReaction := models.NewReaction(models.ReactionConfig{
 		IsLike: cmd.IsLike,
 		UserID: cmd.UserId,
 		BlogID: cmd.BlogId,
 	})
-	if err = h.reactionRepo.Save(newReaction); err != nil {
+
+	if err = h.reactionRepo.Save(*newReaction); err != nil {
 		return false, err
 	}
